Moved changes sim/finitialize.cpp to io/nrn2core_data_init.cpp
